+++ Days/Day02.cs
@@ -2,7 +2,7 @@
 
 internal class Day02 : IDay
 {
-    public async Task<int> FirstPart()
+    public async Task<long> FirstPart()
     {
         using var reader = new StringReader(await File.ReadAllTextAsync("Days/Input/02.txt"));
 
@@ -26,7 +26,7 @@ internal class Day02 : IDay
         return pos.X * pos.Y;
     }
 
-    public async Task<int> SecondPart()
+    public async Task<long> SecondPart()
     {
         using var reader = new StringReader(await File.ReadAllTextAsync("Days/Input/02.txt"));
